- compute re-projection of predicted position ('h') right after the prediction step of the EKF. So it can be used for the tracking. - created a Features_extra structure, which contains all the extra info about the features. 'h', 'H', 'z' and if it is valid. - use dataset from the actual drone - OF is filtering out tracked features outside the image space - OF is now using the predicted information of the EKF as starting point for the OF. - Features which have predicted a negative depth are no longer used.

+++ src/kalman_test.cpp
@@ -312,7 +312,7 @@ void KalmanTestCase::test_delete_features () {
 				1.0    //standar deviation for measurement noise
 		);
 
-	std::vector<int> delete_list;
+	std::vector<size_t> delete_list;
 
 	delete_list.push_back(0);
 	delete_list.push_back(2);
